Move SFT response masking into preprocessing utility
