+++ src/art/preprocessing/tokenize.py
@@ -19,6 +19,7 @@ from ..utils.chat_template import (
     default_chat_template_kwargs_for_tokenizer,
     merge_chat_template_kwargs,
 )
+from .response_masking import response_only_labels, token_ids_for_template_part
 
 ChatTemplateTool = dict[Any, Any] | Callable[..., Any]
 ChatTemplateToolSchemaFormat = Literal["default", "vllm_openai"]
@@ -134,55 +135,6 @@ def _messages_for_chat_template(
                 "content": message.get("content") or "",
             }
     return _normalize_tool_call_arguments_for_chat_template(tokenizer, messages)
-
-
-def _token_ids_for_template_part(
-    tokenizer: PreTrainedTokenizerBase,
-    template_part: str,
-) -> list[int]:
-    return list(tokenizer(template_part, add_special_tokens=False).input_ids)
-
-
-def _find_subsequence(
-    values: list[int],
-    pattern: list[int],
-    *,
-    start: int = 0,
-) -> int | None:
-    if not pattern:
-        return None
-    last_start = len(values) - len(pattern)
-    for index in range(start, last_start + 1):
-        if values[index : index + len(pattern)] == pattern:
-            return index
-    return None
-
-
-def _response_only_labels(
-    input_ids: list[int],
-    *,
-    instruction_ids: list[int],
-    response_ids: list[int],
-) -> list[int]:
-    labels = [-100] * len(input_ids)
-    index = 0
-    while index < len(input_ids):
-        response_start = _find_subsequence(input_ids, response_ids, start=index)
-        if response_start is None:
-            break
-
-        trainable_start = response_start + len(response_ids)
-        next_instruction_start = _find_subsequence(
-            input_ids,
-            instruction_ids,
-            start=trainable_start,
-        )
-        trainable_end = (
-            len(input_ids) if next_instruction_start is None else next_instruction_start
-        )
-        labels[trainable_start:trainable_end] = input_ids[trainable_start:trainable_end]
-        index = trainable_end
-    return labels
 
 
 @dataclass
@@ -632,8 +584,8 @@ def tokenize_sft_batch(
     """
     _validate_max_seq_length(max_seq_length)
 
-    instruction_ids = _token_ids_for_template_part(tokenizer, instruction_part)
-    response_ids = _token_ids_for_template_part(tokenizer, response_part)
+    instruction_ids = token_ids_for_template_part(tokenizer, instruction_part)
+    response_ids = token_ids_for_template_part(tokenizer, response_part)
     # Tokenize all trajectories (no padding — each keeps its natural length)
     trajectory_tensors = []
     num_tokens = 0
@@ -665,7 +617,7 @@ def tokenize_sft_batch(
 
         attention_mask = [1] * len(input_ids)
 
-        labels = _response_only_labels(
+        labels = response_only_labels(
             input_ids,
             instruction_ids=instruction_ids,
             response_ids=response_ids,
