+++ src/art/utils/chat_template.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-THINKING_CHAT_TEMPLATE_KWARGS: dict[str, object] = {
+THINKING_CHAT_TEMPLATE_KWARGS: dict[str, Any] = {
     "enable_thinking": False,
     "preserve_thinking": True,
 }
@@ -8,8 +8,8 @@ THINKING_CHAT_TEMPLATE_KWARGS: dict[str, object] = {
 
 def default_chat_template_kwargs_for_template(
     chat_template: object,
-) -> dict[str, object]:
-    kwargs: dict[str, object] = {}
+) -> dict[str, Any]:
+    kwargs: dict[str, Any] = {}
     if not isinstance(chat_template, str):
         return kwargs
     if "enable_thinking" in chat_template:
@@ -19,14 +19,14 @@ def default_chat_template_kwargs_for_template(
     return kwargs
 
 
-def default_chat_template_kwargs_for_tokenizer(tokenizer: object) -> dict[str, object]:
+def default_chat_template_kwargs_for_tokenizer(tokenizer: object) -> dict[str, Any]:
     return default_chat_template_kwargs_for_template(
         getattr(tokenizer, "chat_template", None)
     )
 
 
 def merge_chat_template_kwargs(
-    defaults: dict[str, object] | None,
+    defaults: dict[str, Any] | None,
     overrides: dict[str, Any] | None,
 ) -> dict[str, Any]:
     return {**(defaults or {}), **(overrides or {})}
